refactor: migrate to React Query for auth state management and remove Zustand dependencies
