--- frontend/src/features/auth/login/api/use-login-v2.tsx
@@ -1,67 +1,0 @@
-"use client";
-
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { LoginCredentials } from "../lib/type";
-import { userQueryKeys } from "@/entities/user/api/auth-hooks";
-
-type LoginMutationOptions = {
-  onSuccess?: (data: unknown, variables: LoginCredentials, context: unknown) => void;
-  onError?: (error: Error, variables: LoginCredentials, context: unknown) => void;
-};
-
-/**
- * Login mutation hook - React Query only version
- * No Zustand dependency - all state managed by React Query
- */
-export const useLoginMutation = ({ onSuccess, onError }: LoginMutationOptions = {}) => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async (credentials: LoginCredentials) => {
-      const response = await fetch(`/api/auth/login`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Accept: "application/json",
-        },
-        body: JSON.stringify(credentials),
-        credentials: "include",
-      });
-
-      if (!response.ok) {
-        const errorData = await response
-          .json()
-          .catch(() => ({ error: "Failed to parse error response" }));
-        throw new Error(errorData.message || errorData.error || "Login failed");
-      }
-
-      return response.json();
-    },
-    onSuccess: (data, variables, ctx) => {
-      // Extract user from response
-      const user = data.data?.user || data.user;
-
-      if (user) {
-        // Update React Query cache with user data
-        queryClient.setQueryData(userQueryKeys.me(), user);
-      }
-
-      // Invalidate to trigger refetch (ensures fresh data)
-      queryClient.invalidateQueries({ queryKey: userQueryKeys.all });
-
-      // Call custom success handler
-      if (onSuccess) {
-        onSuccess(data, variables, ctx);
-      }
-    },
-    onError: (error, variables, ctx) => {
-      console.error("Login error:", error);
-
-      // Call custom error handler
-      if (onError) {
-        onError(error, variables, ctx);
-      }
-    },
-  });
-};
-
